ref, feat: MS Word UL replacement on paste
Refactored to have the logic for creating dom nodes for the editor based on text pasted from ms word apply on paste, rather than on save.

Created custom functions to handle interception and conversion/elements with DOMParser
