+++ src/components/RichTextEditor/Editors/Sections/EditableSRTE.tsx
@@ -9,283 +9,293 @@ import { ListPlugin } from "@lexical/react/LexicalListPlugin";
 import { HistoryPlugin } from "@lexical/react/LexicalHistoryPlugin";
 import { ListItemNode, ListNode } from "@lexical/list";
 import { HeadingNode } from "@lexical/rich-text";
-import { $getRoot, $getSelection, LexicalEditor, NodeKey, ParagraphNode } from "lexical";
+import {
+  $getRoot,
+  $getSelection,
+  LexicalEditor,
+  NodeKey,
+  ParagraphNode,
+} from "lexical";
 import { PrepopulateHTMLPlugin } from "../../Plugins/PrepopulateHTMLPlugin";
 import { RichTextToolbar } from "../../Toolbar/RichTextToolbar";
 import { ContentEditable } from "@lexical/react/LexicalContentEditable";
 import LexicalErrorBoundary from "@lexical/react/LexicalErrorBoundary";
-import { TreeViewPlugin } from "../../../../lib/plugins/TreeViewPlugin"
+import { TreeViewPlugin } from "../../../../lib/plugins/TreeViewPlugin";
 import { useEffect, useState } from "react";
 import { DataDisplayPlugin } from "../../Plugins/DataDisplayPlugin";
-import { $generateNodesFromDOM, $generateHtmlFromNodes } from "@lexical/html"
-import { EditorSections, EditorSubsections, EditorType } from "../../../../types";
+import { $generateNodesFromDOM, $generateHtmlFromNodes } from "@lexical/html";
+import {
+  EditorSections,
+  EditorSubsections,
+  EditorType,
+} from "../../../../types";
 import { useGetRTESectionPlaceholder } from "@/lib/hooks/useGetRTESectionPlaceholder";
-import { AutoFocusPlugin } from '@lexical/react/LexicalAutoFocusPlugin';
-import { TablePlugin } from '@lexical/react/LexicalTablePlugin';
-import DraggableBlockPlugin from '@/components/RichTextEditor/Plugins/DraggableBlockPlugin';
-import { TabIndentationPlugin } from '@lexical/react/LexicalTabIndentationPlugin';
+import { AutoFocusPlugin } from "@lexical/react/LexicalAutoFocusPlugin";
+import { TablePlugin } from "@lexical/react/LexicalTablePlugin";
+import DraggableBlockPlugin from "@/components/RichTextEditor/Plugins/DraggableBlockPlugin";
+import { TabIndentationPlugin } from "@lexical/react/LexicalTabIndentationPlugin";
 import ListMaxIndentLevelPlugin from "../../Plugins/ListMaxIndentLevelPlugin";
 import { EditorTextInitialStatePlugin } from "../../Plugins/EditorTextInitialStatePlugin";
+import { CustomPastePlugin } from "../../Plugins/CustomPastePlugin";
 
 interface Props {
-    initialConfig: any;
-    editorRef: any;
-    data: string;
-    section: EditorSubsections;
-    project_pk: number;
-    document_pk: number;
-    editorType: EditorType;
-    isUpdate: boolean;
-    writeable_document_kind?: EditorSections | null;
-    writeable_document_pk?: number | null;
+  initialConfig: any;
+  editorRef: any;
+  data: string;
+  section: EditorSubsections;
+  project_pk: number;
+  document_pk: number;
+  editorType: EditorType;
+  isUpdate: boolean;
+  writeable_document_kind?: EditorSections | null;
+  writeable_document_pk?: number | null;
 
-    editorText: string;
-    setEditorText: React.Dispatch<React.SetStateAction<string>>;
-    shouldShowTree: boolean;
-    setShouldShowTree: React.Dispatch<React.SetStateAction<boolean>>;
-    isEditorOpen: boolean;
-    setIsEditorOpen: React.Dispatch<React.SetStateAction<boolean>>;
-    displayData: string;
+  editorText: string;
+  setEditorText: React.Dispatch<React.SetStateAction<string>>;
+  shouldShowTree: boolean;
+  setShouldShowTree: React.Dispatch<React.SetStateAction<boolean>>;
+  isEditorOpen: boolean;
+  setIsEditorOpen: React.Dispatch<React.SetStateAction<boolean>>;
+  displayData: string;
 
-    setDisplayData: React.Dispatch<React.SetStateAction<string>>;
-    textEditorName?: string;
-    wordLimit: number;
-    limitCanBePassed: boolean;
+  setDisplayData: React.Dispatch<React.SetStateAction<string>>;
+  textEditorName?: string;
+  wordLimit: number;
+  limitCanBePassed: boolean;
 
-    canSave: boolean;
-    setCanSave: React.Dispatch<React.SetStateAction<boolean>>;
+  canSave: boolean;
+  setCanSave: React.Dispatch<React.SetStateAction<boolean>>;
 }
 
-export const EditableSRTE = (
-    { textEditorName, section, project_pk, document_pk, editorType, isUpdate,
-        writeable_document_kind, writeable_document_pk,
-        initialConfig, editorRef,
-        data, editorText, setEditorText,
-        isEditorOpen, setIsEditorOpen,
-        displayData, setDisplayData,
-        shouldShowTree, setShouldShowTree,
-        wordLimit, limitCanBePassed,
-        canSave, setCanSave,
-    }: Props) => {
+export const EditableSRTE = ({
+  textEditorName,
+  section,
+  project_pk,
+  document_pk,
+  editorType,
+  isUpdate,
+  writeable_document_kind,
+  writeable_document_pk,
+  initialConfig,
+  editorRef,
+  data,
+  editorText,
+  setEditorText,
+  isEditorOpen,
+  setIsEditorOpen,
+  displayData,
+  setDisplayData,
+  shouldShowTree,
+  setShouldShowTree,
+  wordLimit,
+  limitCanBePassed,
+  canSave,
+  setCanSave,
+}: Props) => {
+  const [selectedNodeType, setSelectedNodeType] = useState<string>();
 
-    const [selectedNodeType, setSelectedNodeType] = useState<string>();
+  const [selectedIsBold, setSelectedIsBold] = useState(false);
 
-    const [selectedIsBold, setSelectedIsBold] = useState(false);
+  const dragBtnMargin = 10;
+  // useEffect(() => console.log(displayData), [displayData])
 
+  const toolBarHeight = 42;
+  const [floatingAnchorElem, setFloatingAnchorElem] =
+    useState<HTMLDivElement | null>(null);
+  const onRef = (_floatingAnchorElem: HTMLDivElement) => {
+    if (_floatingAnchorElem !== null) {
+      setFloatingAnchorElem(_floatingAnchorElem);
+    }
+  };
+  useEffect(() => {
+    // Check if floatingAnchorElem is set
+    if (floatingAnchorElem !== null) {
+      // Perform any actions that depend on floatingAnchorElem being set
+      // For example, you might want to trigger something or log a message.
+      console.log("floatingAnchorElem is set:", floatingAnchorElem);
+    }
+  }, [floatingAnchorElem]);
 
+  useEffect(() => {
+    // Assuming editorRef is a reference to your Lexical editor instance
+    const editorInstance = editorRef.current;
 
-    const dragBtnMargin = 10;
-    // useEffect(() => console.log(displayData), [displayData])
+    if (editorInstance) {
+      // Perform operations on the Lexical editor instance
+      const root = editorInstance.$getRoot();
+      setEditorText(root.__cachedText);
+    }
+  }, [editorRef, setEditorText]);
 
-    const toolBarHeight = 42;
-    const [floatingAnchorElem, setFloatingAnchorElem] = useState<HTMLDivElement | null>(null);
-    const onRef = (_floatingAnchorElem: HTMLDivElement) => {
-        if (_floatingAnchorElem !== null) {
-            setFloatingAnchorElem(_floatingAnchorElem);
-        }
-    };
-    useEffect(() => {
-        // Check if floatingAnchorElem is set
-        if (floatingAnchorElem !== null) {
-            // Perform any actions that depend on floatingAnchorElem being set
-            // For example, you might want to trigger something or log a message.
-            console.log('floatingAnchorElem is set:', floatingAnchorElem);
-        }
-    }, [floatingAnchorElem]);
+  return (
+    <>
+      <LexicalComposer initialConfig={initialConfig}>
+        {/* Plugins*/}
+        <TablePlugin hasCellMerge={true} hasCellBackgroundColor={false} />
+        <AutoFocusPlugin />
+        <HistoryPlugin />
+        <ListPlugin />
 
-    useEffect(() => {
-        // Assuming editorRef is a reference to your Lexical editor instance
-        const editorInstance = editorRef.current;
+        <OnChangePlugin
+          onChange={(editorState, editor) => {
+            editorState.read(() => {
+              const root = $getRoot();
+              setEditorText(root.__cachedText);
+              const newHtml = $generateHtmlFromNodes(editor, null);
+              // console.log(newHtml)
+              // console.log(newHtml)
+              // console.log("DATA DISPLAY PLUGIN:", newHtml);
+              setDisplayData(newHtml);
+            });
+          }}
+        />
+        {data !== undefined && data !== null && (
+          <PrepopulateHTMLPlugin data={displayData} />
+        )}
+        <CustomPastePlugin />
 
-        if (editorInstance) {
-            // Perform operations on the Lexical editor instance
-            const root = editorInstance.$getRoot();
-            setEditorText(root.__cachedText);
-        }
-    }, [editorRef, setEditorText]);
-
-
-
-
-    return (
-        <>
-            <LexicalComposer
-                initialConfig={initialConfig}
-
+        {/* Text Area */}
+        <RichTextPlugin
+          contentEditable={
+            <Box
+            // mr={3}
             >
-                {/* Plugins*/}
-                <TablePlugin
-                    hasCellMerge={true}
-                    hasCellBackgroundColor={false}
-                />
-                <AutoFocusPlugin />
-                <HistoryPlugin />
-                <ListPlugin />
+              {/* Toolbar */}
 
-                <OnChangePlugin onChange={(editorState, editor) => {
-                    editorState.read(() => {
-                        const root = $getRoot();
-                        setEditorText(root.__cachedText);
-                        const newHtml = $generateHtmlFromNodes(editor, null)
-                        // console.log(newHtml)
-                        // console.log(newHtml)
-                        // console.log("DATA DISPLAY PLUGIN:", newHtml);
-                        setDisplayData(newHtml);
+              <RichTextToolbar
+                // editor={editor}
+                selectedIsBold={selectedIsBold}
+                setSelectedIsBold={setSelectedIsBold}
+                editorRef={editorRef}
+                selectedNodeType={selectedNodeType}
+                setSelectedNodeType={setSelectedNodeType}
+              />
 
-                    })
-                }} />
-                {data !== undefined && data !== null && (
-                    <PrepopulateHTMLPlugin data={displayData} />
+              <Box className="editor-scroller">
+                <Box
+                  className="editor"
+                  ref={onRef}
+                  style={{
+                    // background: "red",
+                    marginLeft: `${dragBtnMargin}px`,
+                  }}
+                >
+                  <ContentEditable
+                    style={{
+                      minHeight: "50px",
+                      width: "100%",
+                      height: "auto",
+                      padding: "32px",
+                      paddingBottom: "16px",
+                      borderRadius: "0 0 25px 25px",
+                      outline: "none",
+                    }}
 
-                )}
-
-
-                {/* Text Area */}
-                <RichTextPlugin
-                    contentEditable={
-                        <Box
-                        // mr={3}
-                        >
-                            {/* Toolbar */}
-
-                            <RichTextToolbar
-                                // editor={editor}
-                                selectedIsBold={selectedIsBold}
-                                setSelectedIsBold={setSelectedIsBold}
-                                editorRef={editorRef}
-                                selectedNodeType={selectedNodeType}
-                                setSelectedNodeType={setSelectedNodeType}
-                            />
-
-                            <Box className="editor-scroller"
-
-                            >
-                                <Box className="editor" ref={onRef}
-                                    style={{
-                                        // background: "red",
-                                        marginLeft: `${dragBtnMargin}px`,
-                                    }}
-                                >
-                                    <ContentEditable
-                                        style={{
-
-                                            minHeight: "50px",
-                                            width: "100%",
-                                            height: "auto",
-                                            padding: "32px",
-                                            paddingBottom: "16px",
-                                            borderRadius: "0 0 25px 25px",
-                                            outline: "none",
-                                        }}
-
-                                    // autoFocus
-                                    />
-                                </Box>
-                            </Box>
-                            {/* <Box>Editor: {editorText}</Box> */}
-                        </Box>
-                    }
-                    placeholder={
-                        <Box
-                            style={{
-                                position: "absolute",
-                                left: `${32 + dragBtnMargin}px`,
-                                top: `${30 + toolBarHeight}px`,
-                                userSelect: "none",
-                                pointerEvents: "none",
-                                color: "gray"
-
-                            }}
-                        >
-                            {`Enter ${useGetRTESectionPlaceholder(section)}..`}
-                        </Box>
-                    }
-                    ErrorBoundary={LexicalErrorBoundary}
-                />
-                {floatingAnchorElem !== null && (
-                    <DraggableBlockPlugin anchorElem={floatingAnchorElem} toolbarHeight={toolBarHeight} />
-                )}
-
-                <Box>
-                    <OptionsBar
-
-                        editorType={editorType}
-                        displayData={displayData}
-                        editorText={editorText}
-                        shouldShowTree={shouldShowTree}
-                        setShouldShowTree={setShouldShowTree}
-                        rawHTML={displayData}
-                        section={section}
-                        project_pk={project_pk}
-                        document_pk={document_pk}
-                        isUpdate={isUpdate}
-                        writeable_document_kind={writeable_document_kind}
-                        writeable_document_pk={writeable_document_pk}
-                        wordLimit={wordLimit}
-                        limitCanBePassed={limitCanBePassed}
-
-                        canSave={canSave}
-                        setCanSave={setCanSave}
-
-                        editorIsOpen={isEditorOpen}
-                        setIsEditorOpen={setIsEditorOpen}
-                        setDisplayData={setDisplayData}
-                    />
+                    // autoFocus
+                  />
                 </Box>
-                {
-                    shouldShowTree ?
-                        <TreeViewPlugin />
-                        : null
-                }
-                <ClearEditorPlugin />
-                <TabIndentationPlugin />
-                <ListMaxIndentLevelPlugin maxDepth={3} />
+              </Box>
+              {/* <Box>Editor: {editorText}</Box> */}
+            </Box>
+          }
+          placeholder={
+            <Box
+              style={{
+                position: "absolute",
+                left: `${32 + dragBtnMargin}px`,
+                top: `${30 + toolBarHeight}px`,
+                userSelect: "none",
+                pointerEvents: "none",
+                color: "gray",
+              }}
+            >
+              {`Enter ${useGetRTESectionPlaceholder(section)}..`}
+            </Box>
+          }
+          ErrorBoundary={LexicalErrorBoundary}
+        />
+        {floatingAnchorElem !== null && (
+          <DraggableBlockPlugin
+            anchorElem={floatingAnchorElem}
+            toolbarHeight={toolBarHeight}
+          />
+        )}
 
-                <NodeEventPlugin
-                    nodeType={ParagraphNode}
-                    eventType={'click'}
-                    eventListener={(e: Event, editor: LexicalEditor, nodeKey: NodeKey) => {
-                        // editor.getEditorState().read(() => {
-                        //     // 
-                        //     setSelectedIsBold(true);
-                        // })
-                        console.log(e)
-                        console.log('paragaph node clicked')
-                        setSelectedNodeType('Normal')
-                        setSelectedIsBold(true);
+        <Box>
+          <OptionsBar
+            editorType={editorType}
+            displayData={displayData}
+            editorText={editorText}
+            shouldShowTree={shouldShowTree}
+            setShouldShowTree={setShouldShowTree}
+            rawHTML={displayData}
+            section={section}
+            project_pk={project_pk}
+            document_pk={document_pk}
+            isUpdate={isUpdate}
+            writeable_document_kind={writeable_document_kind}
+            writeable_document_pk={writeable_document_pk}
+            wordLimit={wordLimit}
+            limitCanBePassed={limitCanBePassed}
+            canSave={canSave}
+            setCanSave={setCanSave}
+            editorIsOpen={isEditorOpen}
+            setIsEditorOpen={setIsEditorOpen}
+            setDisplayData={setDisplayData}
+          />
+        </Box>
+        {shouldShowTree ? <TreeViewPlugin /> : null}
+        <ClearEditorPlugin />
+        <TabIndentationPlugin />
+        <ListMaxIndentLevelPlugin maxDepth={3} />
 
-                    }}
-                />
-                <NodeEventPlugin
+        <NodeEventPlugin
+          nodeType={ParagraphNode}
+          eventType={"click"}
+          eventListener={(
+            e: Event,
+            editor: LexicalEditor,
+            nodeKey: NodeKey
+          ) => {
+            // editor.getEditorState().read(() => {
+            //     //
+            //     setSelectedIsBold(true);
+            // })
+            console.log(e);
+            console.log("paragaph node clicked");
+            setSelectedNodeType("Normal");
+            setSelectedIsBold(true);
+          }}
+        />
+        <NodeEventPlugin
+          nodeType={ListItemNode}
+          eventType={"click"}
+          eventListener={(
+            e: Event,
+            editor: LexicalEditor,
+            nodeKey: NodeKey
+          ) => {
+            console.log(e);
+            console.log(editor);
+            console.log(nodeKey);
+            const a = false;
+            if (a) {
+              setSelectedNodeType("Bullets");
+            } else {
+              setSelectedNodeType("Nums");
+            }
 
-                    nodeType={ListItemNode}
-                    eventType={'click'}
-                    eventListener={(e: Event, editor: LexicalEditor, nodeKey: NodeKey) => {
-
-                        console.log(e)
-                        console.log(editor)
-                        console.log(nodeKey)
-                        const a = false;
-                        if (a) {
-                            setSelectedNodeType('Bullets')
-
-                        } else {
-                            setSelectedNodeType("Nums")
-                        }
-
-                        // console.log('li node clicked')
-                        // setSelectedNodeType('li')
-
-                    }}
-                />
-                <EditorTextInitialStatePlugin
-                    setEditorText={setEditorText}
-                    editorText={editorText}
-                    isOpen={isEditorOpen}
-
-                />
-                {/* <DataDisplayPlugin setDisplayData={setDisplayData} /> */}
-                {/* {isEditorOpen && (
+            // console.log('li node clicked')
+            // setSelectedNodeType('li')
+          }}
+        />
+        <EditorTextInitialStatePlugin
+          setEditorText={setEditorText}
+          editorText={editorText}
+          isOpen={isEditorOpen}
+        />
+        {/* <DataDisplayPlugin setDisplayData={setDisplayData} /> */}
+        {/* {isEditorOpen && (
                     <EditorTextInitialStatePlugin
                         setEditorText={setEditorText}
                         editorText={editorText}
@@ -293,10 +303,7 @@ export const EditableSRTE = (
                     />
 
                 )} */}
-
-            </LexicalComposer>
-        </>
-    )
-}
-
-
+      </LexicalComposer>
+    </>
+  );
+};
